% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{collapse_RowsUpIf}
\alias{collapse_RowsUpIf}
\title{Merge Multiple Rows that Contain Data on the Same Subscriber}
\usage{
collapse_RowsUpIf(
  DTname,
  vCol,
  Regex_str1,
  weight1 = 1,
  Regex_str2 = NA,
  Regex_str3 = NA,
  nchar_from_front = -1,
  nchar_from_back = -1
)
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{vCol}{Character. Name of the column within the R object to be operated on.}

\item{Regex_str1}{Character. Regular Expression (RegEx) pattern to be searched for within
character strings and, following that, to apply the operation onto.}

\item{weight1}{Numeric - optional. Weighting of the RegEx pattern \code{Regex_str1} in determining
whether or not to collapse a row of records to the row above. The higher this number, the more
important \code{Regex_str1} will be in comparison to the other two RegEx patterns. Default value \verb{=1}.}

\item{Regex_str2}{Character - optional. See \code{Regex_str1}.}

\item{Regex_str3}{Character - optional. See \code{Regex_str1}.}

\item{nchar_from_front}{Numeric - optional. Number of characters from the front of
the character string to restrict the string search operation to. If this argument is used,
then the argument \code{nchar_from_back} should not be used.}

\item{nchar_from_back}{Numeric - optional. Number of characters from the back of
the character string to restrict the string search operation to. If this argument is used,
then the argument \code{nchar_from_front} should not be used.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column(s).
}
\description{
This function combines consecutive rows that contain information on the same telephone subscriber
by using up to three different Regular Expressions (RegEx) string pattern queries. Based on whether
the patterns in these queries are detected (indicating a split-up record), every row of data
will be given a score and merged into the row above it if their score is higher than that of the other row.
}
