% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ocr_text.R
\name{create_emptyDT}
\alias{create_emptyDT}
\title{Creates Empty Object to Store All Output Text from Telephone Directories}
\usage{
create_emptyDT(masterDTname = "dtAPH", num_vCol = 3, pg_split = TRUE)
}
\arguments{
\item{masterDTname}{Character. Name of the empty object to be created in the global environment.}

\item{num_vCol}{Numeric. Number of "V-" Columns to create in the new data frame/table object.
These will be used to store information such as telephone number, subscriber name and address.}

\item{pg_split}{Logical. Did one page of the telephone directories concerned comprise
more than one column of data? If yes, an additional column \code{PgCol} will be made. Default \verb{=TRUE}.}
}
\value{
An empty table of both types 'data.table' and 'data.frame', to which information can be appended.
}
\description{
This function initialises an empty object in the global environment to store ALL the processed
data from the outputs of Optical Character Recognition (OCR) from one directory. This is opposed to
some other functions in the package which deal specifically with the outputs from a single .txt. file.
}
