% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsequent_data_prep.R
\name{create_newCols}
\alias{create_newCols}
\title{Creates One or More Empty Columns}
\usage{
create_newCols(DTname = "tld", newcols = c("tld_num", "tld_street"))
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{newcols}{Vector of character strings. Names of empty column(s) to be created, of type 'character'.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column(s).
}
\description{
This function creates one or more empty columns in a given data table or data frame.
}
