% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{del_orCollapsebyNA}
\alias{del_orCollapsebyNA}
\title{Delete Rows and/or Collapse Information across Rows using 'NA's}
\usage{
del_orCollapsebyNA(
  DTname,
  vColRange,
  origCol = NA,
  addtoCol = NA,
  export_thres = 1
)
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{vColRange}{Character or Numeric Range. If character, this refers to the name of the column
within the R object to be operated on. If a numeric range, this refers to the range of 'V-Columns'}

\item{origCol}{Character - optional. If desired, the name of the column from which any character
string remaining in a row slated for deletion is to be transferred to the row above.}

\item{addtoCol}{Character - optional. If desired, the name of the column to which any character
string remaining in a row slated for deletion is to be transferred from the row below.}

\item{export_thres}{Integer. Number of permissible \code{'NA'} values per row of subscriber records,
out of the cells in the columns listed in \code{vColRange}. Default value \verb{=1}.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column(s).
}
\description{
This function either deletes (by default) or collapses rows into the rows above them,
conditional on the number of \code{'NA'} values that each row has among the columns specified in
\code{vColRange}. In the case of deletion, rows must have a number of columns not holding \code{'NA'} values
corresponding to the value of \code{export_thres} to not be deleted. In the case of collapsing of rows,
}
