% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsequent_data_prep.R
\name{detect_isBusy}
\alias{detect_isBusy}
\title{Distinguish Business and Residential Subscribers}
\usage{
detect_isBusy(DTname = "tld", vCol)
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{vCol}{Character. Name of the column within the R object to be operated on.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column(s).
}
\description{
This function creates a new column \code{isBusiness}, if not already existing. It then uses a list
of pre-defined strings stored in the package to infer if a particular telephone directory record
belongs to a commercial (business) or personal (residential) subscriber.
}
