% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_addr.R
\name{export_fullDT}
\alias{export_fullDT}
\title{Exports a Data Table}
\usage{
export_fullDT(
  DTname = "dtAPH",
  dirYR = NULL,
  xpNAME = ".Ls.Adr.",
  csv = 1,
  xlsx = 0,
  xpSUBDIR = "1.Addr.Tables"
)
}
\arguments{
\item{DTname}{Character. Name of the object in the global environment to be exported.}

\item{dirYR}{Numeric. Year about which the table to be exported holds information. This argument will
be used to place exports automatically in a sub-folder corresponding to the year. The default is
\verb{=NULL}, in which case the year information will be automatically taken from the table.}

\item{xpNAME}{Character. Forms part of the name of the outputted file(s). This will be
combined with a timestamp of when the file was exported.}

\item{csv}{Logical. If \verb{=1}, the object will be exported as a \code{.csv} file. Default is \verb{=1}.}

\item{xlsx}{Logical. If \verb{=1}, the object will be exported as a \code{.xlsx} file. Default is \verb{=0}.}

\item{xpSUBDIR}{Character. A sub-directory to which the object will be exported.
It is nestled within the file path: "2.Intermediary/" and the default sub-directory
therein is "1.Addr.Tables".}

\item{FLname}{Character. Name of the table to be imported, including the file extension.
This can either be a \code{.csv} or an \code{.xslx} file.}
}
\value{
An object that is both of class 'data.table' and 'data.frame'.
}
\description{
This function loads a processed table of telephone subscriber addresses as an object of classes
'data.table' and 'data.frame' from a given file path. It automatically determines the sub-path
to the file (i.e. the year of which it holds information) and allows for users to choose a random
subset of addresses to trial their analysis on.
}
