% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{fillCells_withNA}
\alias{fillCells_withNA}
\title{Conditionally Fill Cells with 'NA' Values}
\usage{
fillCells_withNA(DTname, vColRange, nchar_thres = 3, Remove_str = NA)
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{vColRange}{Character or Numeric Range. If character, this refers to the name of the column
within the R object to be operated on. If a numeric range, this refers to the range of 'V-Columns'
that will be operated on, e.g. \code{2:5} refers to the columns \code{c("V2", "V3", "V4", "V5")}.}

\item{nchar_thres}{Numeric. If the number of characters in a string in the specified column range
is less than or equal to this number, the values inside them will be replaced with \code{'NA'}s. If
instead this value is negative, then \code{'NA'}s will replace cells whose character lengths exceed
the absolute value of this number. Default value is \verb{=3}.}

\item{Remove_str}{Character - optional. Regular Expressions (RegEx) string pattern to search
for and remove before counting the number of characters in each string in the chosen column. This
value is blank by default, meaning no character in any string will be disregarded.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column(s).
}
\description{
This function replaces values within a column that are likely to not contain useful information:
they fulfil a condition relating to the length of string, i.e. number of characters, as specified
by the user. By default, any string values of length \code{nchar_thres} or \emph{lower} get replaced with \code{'NA'}
values. However, if \code{nchar_thres} is a negative number, any strings of length \emph{above} the absolute
value of \code{nchar_thres} will be replaced with \code{'NA'} instead.
}
