% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lstABRV}
\alias{lstABRV}
\title{List of Abbreviations of Throughfare Names}
\format{
\subsection{\code{lstABRV}}{

A single character string of throughfare abbreviations, separated by pipes \code{|}.
}
}
}
\source{
\if{html}{\out{
<Compiled by creators of the teldiR Package>
}}
}
\usage{
lstABRV
}
\description{
A Regular Expressions (RegEx) pattern for detecting the presence of throughfare
names in their abbreviated forms.
}
\keyword{datasets}
