% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lstTOPO}
\alias{lstTOPO}
\title{List of Abbreviations of Throughfare Names}
\format{
\subsection{\code{lstTOPO}}{

A single character string of throughfare name, separated by pipes \code{|}.
}
}
}
\source{
\if{html}{\out{
<Compiled by creators of the teldiR Package>
}}
}
\usage{
lstTOPO
}
\description{
A Regular Expressions (RegEx) pattern for detecting the presence of throughfare
names in their full (non-abbreviated) forms.
}
\keyword{datasets}
