% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsequent_data_prep.R
\name{proofcheck_Surnames}
\alias{proofcheck_Surnames}
\title{Checks and Corrects List of Surnames}
\usage{
proofcheck_Surnames(DTname = "tld", DTrefName = "srn", maxDist = 1)
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{DTrefName}{Character. Name of the R object (a table) which serves as the reference table,
against which surnames from \code{DTname} are to be fuzzy matched.}

\item{maxDist}{Numeric - optional. Maximum string distance between two character strings (of surnames)
for them to be considered a match to one another.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column.
}
\description{
N.B. This currently works with \verb{\%do\%} in the package but not with \verb{\%dopar\%} (which only works
outside the package, if the function is loaded into global environment!)
This function matches uses fuzzy string matching surnames contained in  one column of a data frame
or data table against another list of (standardised) surnames that the user provides. If a surname
from the former data table is found to closely resemble one from the latter table (as defined by
\code{maxDist}), the value from the latter is appended to the former.
}
