% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_addr.R
\name{readin_adrDT}
\alias{readin_adrDT}
\title{Imports a Data Table of Telephone Subscriber Addresses}
\usage{
readin_adrDT(
  FLname,
  IMPORTname = "tld",
  sub_rows = -1,
  xpSUBDIR = "1.Addr.Tables"
)
}
\arguments{
\item{FLname}{Character. Name of the table to be imported, including the file extension.
This can either be a \code{.csv} or an \code{.xslx} file.}

\item{IMPORTname}{Character. Name of the object to be imported into the global environment.}

\item{sub_rows}{Integer. Number of rows of records to randomly subset from the original
list, which allows testing of code on a smaller number of records. The default value \verb{=-1}
and means no subsetting will take place.}

\item{xpSUBDIR}{Character. A sub-directory from which the table of addresses is to be taken.
It is nestled within the file path: "2.Intermediary/" and the default sub-directory
therein is "1.Addr.Tables".}
}
\value{
An object that is both of class 'data.table' and 'data.frame'.
}
\description{
This function loads a processed table of telephone subscriber addresses as an object of classes
'data.table' and 'data.frame' from a given file path. It automatically determines the sub-path
to the file (i.e. the year of which it holds information) and allows for users to choose a random
subset of addresses to trial their analysis on.
}
