% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ocr_text.R
\name{readin_PG}
\alias{readin_PG}
\title{Imports Text from a Single Text File}
\usage{
readin_PG(fl)
}
\arguments{
\item{fl}{Numeric. Index or position of the name of file to be imported, in the object
\code{ls_aphpg} that was created using the function \code{gen_fileLST()}.}
}
\value{
A table of two columns, containing information from a single .txt OCR output file.
}
\description{
This function imports the information from a text file, usually the outputs of an OCR
algorithm and structures it into a table in preparation for subsequent operations.
}
