% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{repl_Dittos}
\alias{repl_Dittos}
\title{Replaces 'ditto' Strings and Equivalents}
\usage{
repl_Dittos(
  DTname,
  vCol,
  nchar_from_front = 6,
  Surname_str = "^.{0,2}[A-Z]{3,}(?=\\\\s)",
  Ditto_str = "d[o\\\\.\\\\,]{2,4}|(d[o|\\\\.]\\\\s)"
)
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{vCol}{Character. Name of the column within the R object to be operated on.}

\item{nchar_from_front}{Numeric - optional. Number of characters from the front of
the character string to restrict the string operation to. By default the value is \verb{=6}.}

\item{Surname_str}{Character. First Regular Expression (RegEx) pattern to be searched for within
the chosen column \code{vCol}. This RegEx pattern should correspond to how the directory records
the given surname (or forename) of subscribers. By default the string is \code{"^.{0,2}[A-Z]{3,}(?=\\\\s)"}.}

\item{Ditto_str}{Character. Second Regular Expression (RegEx) pattern to be searched for within
the chosen column \code{vCol}. This RegEx pattern should correspond to the character strings that
possibly indicate the use of 'ditto'. By default the string is \code{"d[o\\\\.\\\\,]{2,4}|(d[o|\\\\.]\\\\s)"}.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column.
}
\description{
This function searches for instances of 'ditto' or equivalent character strings in a given column
and replaces them with a surname from the most proximal row above the row with a 'ditto' string.
}
