% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsequent_data_prep.R
\name{repl_streetAbbrevs}
\alias{repl_streetAbbrevs}
\title{Replace Abbreviations of Throughfare Names with their Long-forms}
\usage{
repl_streetAbbrevs(DTname = "tld", vCol, Regex_allowance = "\\\\s?")
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{vCol}{Character. Name of the column within the R object to be operated on.}

\item{Regex_allowance}{Character - optional. Regular Expression (RegEx) pattern allowance for
character strings that are allowed to exist after the abbreviated throughfare names, but before
the end of the string. By default, the allowance is a single space.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column.
}
\description{
This function searches for abbreviated forms of throughfare names in a column containing street
addresses (e.g. lane, street, road...) and replaces them with their long-forms. By default this
function searches for them at the end of character strings, with an optional allowance of a RegEx
string, \code{Regex_allowance}, that the user can choose.
}
