% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{replOrPurge_Col}
\alias{replOrPurge_Col}
\title{Replace or Remove String Pattern in Column(s) of Data}
\usage{
replOrPurge_Col(DTname, vCol1, Regex_str, replOrPurge_str = "", vCol2 = NA)
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{Regex_str}{Character. Regular Expression (RegEx) pattern to be searched for within
character strings and, following that, to apply the operation onto.}

\item{replOrPurge_str}{Character - optional. The string with which any instances of matches of
the RegEx pattern specified in \code{Regex_str} should be replaced. The default value is blank,
meaning that all instances of matches should be deleted.}

\item{vCol2}{Character - optional. If desired, name of the second column within the R object
to be operated on in the same way as \code{vCol}.}

\item{vCol}{Character. Name of the column within the R object to be operated on.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column(s).
}
\description{
This function either replaces all instances of, or completely removes (the default), a given
Regular Expressions (RegEx) string pattern query in one or two selected columns.
}
