% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{split_strIfNA}
\alias{split_strIfNA}
\title{Split Character Strings if Record is Missing a Value}
\usage{
split_strIfNA(
  DTname,
  ifNACol,
  fromCol,
  toCol,
  Regex_str,
  rvrs_frontback = -1,
  nchar_from_front = -1,
  nchar_from_back = -1
)
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{ifNACol}{Character. Name of the column which should be checked for \code{'NA'} values. The splitting
of string will only happen conditional on the value in this column for a record being \code{'NA'}.}

\item{fromCol}{Character. Name of the column from which the string to be split is taken.}

\item{toCol}{Character. Name of the column to which the split-off string will be allocated.}

\item{Regex_str}{Character. Regular Expression (RegEx) pattern to be searched for within
character strings and, following that, to apply the operation onto.}

\item{rvrs_frontback}{Binary - optional. If \verb{=1}, then the splitting operation is reversed such that the
string segment following the matched RegEx pattern in \code{Regex_str} is retained in the column
\code{fromCol} while the string segment preceding that gets moved to the column \code{toCol}.}

\item{nchar_from_front}{Numeric - optional. Number of characters from the front of
the character string to restrict the string operation to. If this argument is used,
then the argument \code{nchar_from_back} should not be used.}

\item{nchar_from_back}{Numeric - optional. Number of characters from the back of
the character string to restrict the string operation to. If this argument is used,
then the argument \code{nchar_from_front} should not be used.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column(s).
}
\description{
This function, like \code{split_strings()}, separates the contents of a character string
into two columns using a Regular Expressions (RegEx) string pattern query. The difference
is that \code{split_strIfNA()} will do this only on select rows of data/records, conditional on
the missingness (\code{'NA'} value) in another specified column.
}
