% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{split_strings}
\alias{split_strings}
\title{Split Character Strings into Two Columns}
\usage{
split_strings(
  DTname,
  fromCol,
  toCol,
  Regex_str,
  nchar_from_front = -1,
  nchar_from_back = -1,
  rvrs_frontback = -1,
  extract_front_only = 0,
  extract_back_only = 0
)
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{fromCol}{Character. Name of the column from which the string to be split is taken.}

\item{toCol}{Character. Name of the column to which the split-off string will be allocated.}

\item{Regex_str}{Character. Regular Expression (RegEx) pattern to be searched for within
character strings and, following that, to apply the operation onto.}

\item{nchar_from_front}{Numeric - optional. Number of characters from the front of
the character string to restrict the string operation to. If this argument is used,
then the argument \code{nchar_from_back} should not be used.}

\item{nchar_from_back}{Numeric - optional. Number of characters from the back of
the character string to restrict the string operation to. If this argument is used,
then the argument \code{nchar_from_front} should not be used.}

\item{rvrs_frontback}{Binary - optional. If \verb{=1}, then the splitting operation is reversed such that the
string segment following the matched RegEx pattern in \code{Regex_str} is retained in the column
\code{fromCol} while the string segment preceding that gets moved to the column \code{toCol}. If this
argument is used, then \code{extract_front_only} and \code{extract_back_only} should not be used.}

\item{extract_front_only}{Binary - optional. If \verb{=1}, then only string segment preceding the matched
RegEx pattern in \code{Regex_str} is copied to the column \code{toCol}, while the column \code{fromCol}. If this
argument is used, then \code{rvrs_frontback} and \code{extract_back_only} should not be used.
remains untouched.}

\item{extract_back_only}{Binary - optional. If \verb{=1}, then only string segment following the matched
RegEx pattern in \code{Regex_str} is copied to the column \code{toCol}, while the column \code{fromCol}. If this
argument is used, then \code{rvrs_frontback} and \code{extract_front_only} should not be used.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column(s).
}
\description{
This function separates the contents of a character string into two columns using
a Regular Expressions (RegEx) string pattern query.
}
