% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{trim_string}
\alias{trim_string}
\title{Remove Character Strings of Unwanted Characters}
\usage{
trim_string(
  DTname = "pgAPH",
  vCol,
  Regex_str,
  nchar_from_back = -1,
  nchar_from_front = -1
)
}
\arguments{
\item{DTname}{Character. Name of the R object (a table) on which the operation is to be run.}

\item{vCol}{Character. Name of the column within the R object to be operated on.}

\item{Regex_str}{Character. Regular Expression (RegEx) pattern to be searched for within
character strings and, following that, to apply the operation onto.}

\item{nchar_from_back}{Numeric - optional. Number of characters from the back of
the character string to restrict the string operation to. If this argument is used,
then the argument \code{nchar_from_front} should not be used.}

\item{nchar_from_front}{Numeric - optional. Number of characters from the front of
the character string to restrict the string operation to. If this argument is used,
then the argument \code{nchar_from_back} should not be used.}
}
\value{
The same object in the global environment (a table) after the string operation
has been applied to the relevant column(s).
}
\description{
This function trims unwanted characters of strings in a particular column at two ends - either
the back (by default) or the front -, using a Regular Expressions (RegEx) string pattern query.
}
