% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_addr.R
\name{view_DTif}
\alias{view_DTif}
\title{View a Subset of the Data}
\usage{
view_DTif(DTname, ifCol1, ifCol2 = NA, nchar_thresh = -1)
}
\arguments{
\item{DTname}{Character. Name of the object in the global environment to be exported.}

\item{ifCol1}{Character. Name of the first column within the R object to check the condition specified
in \code{nchar_thresh} on.}

\item{ifCol2}{Character - optional. Name of the second column within the R object to check the condition on.}

\item{nchar_thresh}{Numeric - optional. This value determines which rows of the table \code{DTname} will be shown
using \code{View()}, whereby only records with values in the columns \code{ifCol1} and \code{ifCol2} (where applicable)
with a string length equal to or lower than this value will be displayed. If this argument is not specified,
then any rows with \code{'NA'} values in the chosen columns will be displayed.}
}
\value{
A view of the same object (a table) but only including rows that fulfil the condition.
}
\description{
This function helps users troubleshoot possible errors or oversights in the text cleaning and
processing by displaying a subset of the data according to the condition defined by the user.
By default, all rows with \code{'NA'} values in the column \code{'ifCol1'} are displayed. If the \code{nchar_thresh}
argument is used, then instead, rows with strings of length equal to or lower than this threshold
are shown, allowing users to diagnose what the issue might be that led to unusually short character strings.
}
