% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ocr_text.R
\name{readin_PG}
\alias{readin_PG}
\title{Imports Text from a Single Text File}
\usage{
readin_PG(fl, sep = "+")
}
\arguments{
\item{fl}{Numeric. Index or position of the name of file to be imported, in the object
\code{ls_aphpg} that was created using the function \code{gen_fileLST()}.}

\item{sep}{Character. Single-byte character that is treated as the marker of separation
of fields within a single line (row). The default value is \code{"+"} which is assumed
to be a character that will not appear in any of the input text files, and this is
desired so that there will not yet be any separation of information into fields in the output.}
}
\value{
A table of two columns, containing information from a single .txt OCR output file.
}
\description{
This function imports the information from a text file, usually the outputs of an OCR
algorithm and structures it into a table in preparation for subsequent operations.
}
